Initial CBOR Encoder/Decoder structure
- Integrated TinyCBOR to be used as the CBOR encoder/decoder.
- Updated SdoBlockio to use TinyCBOR to encode/decode messages
to-and-fro CBOR. The structures have been updated to use
CborEncoder/CborValue and it resembles a Doubly-linked-list to chain
the TinyCBOR's CborEncoder and CborValue structures.
Additionally, added a Unit-test for the same.
- Updated the lib/*.c classes to remove the old methods. Also, excluded
the lib/prot/*.c from compilation for now. The implementation for
changes will happen during FIDOIOT Protocol Spec's
implementation. TO-DO comments have been added for the same at multiple
places. Excluded the same classes from the test cases.

Signed-off-by: Chandrakar, Prateek <prateek.chandrakar@intel.com>