Transfer Ownership 2 (TO2) implementation in CBOR
- Updating TO2 protocol messages (60-71) as per the FDO spec.
As a part of implementation, various methods and structures
for handling Encrypted Message of Composed (EncThenMacMessage) type
have been added. The handling of Encrypted Message Simple type has not
been added yet (to be done at a later time), and the filenames for
messages have not yet been updated for ease in seeing diffs.
- Serviceinfo handling is not a part of this implementation, which will
be done separately.
- Additionally, the Retry scenario has been updated to reflect the
following pseudo-loop:
for-each (retry-counter) {
 for-each (RVDirectives for TO1) {
  if (RVBypass) {
    Run TO2
  }
  for-each (RVTO2Addr) {
    Run TO2
  }
 }
}
- Few memory leaks have been fixed.

NOTE: There are places where the implementation is as per PRI with minor
differences from FDO spec. These will be addressed once PRI is updated.

Signed-off-by: Chandrakar, Prateek <prateek.chandrakar@intel.com>