Fix for issues in rvbypass and TO2 TLS flow (#24)
- Fixing an issue where RVBYPASS is not working

Fixing the RVBYPASS flow and updating the bool structure being used for
it. Subsequently, the bool wrapper struct has been removed.

- Fixing an issue where TO2 is not started with TLS

Fixing an issue where TO2 was not starting with TLS in RVBYPASS flow
when its been specified in RendezvousInfo by passing the TLS flag when
found. However, this fixes the TO2 TLS flow for RVTO2Addr as well.

- The protocol allows use of both RVPROTTLS and RVPROTHTTPS to specify
TLS connection. Updated source to accept both the values to initiate TLS
connection.

Signed-off-by: Chandrakar, Prateek <prateek.chandrakar@intel.com>