Fix return type of fdo_blob_size
Previously, fdo_blob_size() was returning values of type int32_t, but
the return value was used as size_t in many places.

Since the typical return type of the size of a file is size_t, the
return type of fdo_blob_size() was changed to size_t.

This required following additional changes.
1. Update fdo_blob_size() to send 0 for error conditions.
2. The check on return value of fdo_blob_size() was updated to align
with this change.

Signed-off-by: Behera, Tushar Ranjan <tushar.ranjan.behera@intel.com>