Cleanup/organize build/source flags and remove DH Key exchange
- Removed the build flags: KEX, MANUFACTURER_TOOLKIT, MODULES and
PK_ENC. Certain flags in the source that are dependent on the listed
flags have been either removed (MODULES_DEFINED) or set at
alternate locations (KEX_*, AES_*).
- The flags associated with crypto operations are moved into a separate
header file 'crypto/fdoCryptoCommons.h'. This file is then included in
'crypto/fdoCrypto.h' and 'crypto/fdoCryptoHal.h'. As a part of this, the
'DA' flag detrmines the crypto strength of the Key-Exchange, AES and
Hash/HMAC operations, as specified in Section 3.6 of FDO specification.
- READMEs have been updated to reflect the above changes.

Additionally, the Diffie-Hellman Key-Exchange support has been removed.

Signed-off-by: Chandrakar, Prateek <prateek.chandrakar@intel.com>