Update Device ServiceInfo processing and separate buffer size from MTU
- Split the Device ServiceInfo (devmod + unsupported module list) based
on the MTU as specified by the Owner and supported by Device, and send
across multiple msg/68-69 iterations.
- The splitting is done by considering an additional margin for CBOR
encoding. The key-values are CBOR encoded once to decide how many
key-value pairs (partial/complete), can be fitted within the
current message as per MTU. Based on the findings, the ServiceInfo is
CBOR encoded again. This is done in this way, since the underlying
TinyCBOR library doesn't allow us to change the total number of entries
in an array (ServiceInfoKeyVal, in this case), once it's set. As result,
we find the total number of array entries first, by first round of CBOR
encoding + partial/complete key-value entries, and then do the 2nd round
of CBOR encoding.

- Additionally, lowering the range for ServiceInfo size from
[1300-8192] to [256-8192]. As a result, the protocol buffer size
has been separated from maximum ServiceInfo sizes, since the
minimum supported value (256 bytes)+margin,
may not be enough to send protocol messages.

- Fixing an issue in 'main.c' where 'malloc' was done instead of
'fdo_alloc' that caused random values to be present in the structure
allocated.

Signed-off-by: Chandrakar, Prateek <prateek.chandrakar@intel.com>