Update COSE/EAT Signatures to contain Sig_structure signature
- Update the COSE/EAT structures to sign Sig_Structure(RFC 8152),
instead of the payload being signed previously. The Sig_structures is
being generated in separate methods for COSE/EAT, so as to avoid
resetting the 'fdow' objects at the message/types classes.
- The sign/verify Openssl implementations have been updated to
generate/verify signatures of the form 'r|s' that is 'r' concatenated to
's', instead of processing DER encoded signatures previously.

Signed-off-by: Chandrakar, Prateek <prateek.chandrakar@intel.com>