Fix COSE unprotected header parsing
- Previously, the COSE unprotected header map was being parsed in an
ordered manner (entry1, entry2 and so on..). Update the parsing to read
the same map for the same elemnts in an iterative, un-ordered manner.
The header is only supposed to either contain 0 or 2 specific entries.
For any unknown entry, parsing would return an error.
NOTE: Similar parsing be done for other maps as well. However, as of
now, they only contain 1 single entry. So these are being parsed in an
ordered manner (unchanged for now). If more elements need to be parsed,
then those would need to be updated.

- Additionally, fix certain commenting issues, add few NULL checks and
remove unnecessary cose-free being done.

Signed-off-by: Chandrakar, Prateek <prateek.chandrakar@intel.com>