Move new hmac key storage to msg/70 and devmod to TO2 init
- The replacement HMAC key as generated in msg/66 was also being
committed into internal memory, replacing the original HMAC key. This
caused the HMAC verification to fail in msg/61 in case TO2 was being
re-tried for the n'th (1-5) time. Instead of re-loading the same at TO2
init, created new methods to separate the current HMAC key and
replacement HMAC key, and only 'commit' replacement HMAC key at msg/70,
if reuse was not triggered (with reuse, it is never initialized). This
allows futher implementations to manage the replacement HMAC key in
their own way (for example, TPM stores different objects files, instead
of replacing the original).

- Move the 'devmod' allocation and key-value pais creation to TO2 init,
from app init. This fixes an issue wherein the 'devmod' was not being
sent in case initial TO2 failed after completion of msg/68 and
TO2 was being re-tried.

Signed-off-by: Chandrakar, Prateek <prateek.chandrakar@intel.com>