Add more stringent checks for manufacturer_addr.bin
Parse the contents of manufacturer_addr.bin manually for the format
{http,https}://<DNS/IP>:port. The parsing is done as follows with all
the requirements mandatory:
- The transport protocol value will be either `http` or `https`
(case-sensitive). Any other value will result in an error.
- The DNS/IP string's maximum value is 100 characters and must only contain
alphanumeric characters (0-9A-za-z), hyphens (-) and dot (.). First
parsed as IP. If it fails, then treated as DNS.
- The port should be an integer between (1-65535).
- The URL separators are `://` and `:`. These should never be
out-of-place.

NOTE: This does not do a full URL validation as per the URL RFC standards.

Signed-off-by: Chandrakar, Prateek <prateek.chandrakar@intel.com>