Disable setting "Expect 100" in request header when request data size is larger than 1024 bytes
Curl automatically sets it when the request is a POST and the data size is larger than 1024 bytes.
Requests with the Expect: 100-Continue header have an increased probability of becoming separated from one another,
and hence returning with an error.
It can be disabled via setting the Expect: header to the empty string.

Signed-off-by: Shrikant Temburwar <shrikant.temburwar@intel.com>